requirements.txt
Django==1.8.4
PyYAML==3.11
